benchmark/requirements.txt
aiohttp>=3.9.0
pandas>=2.0.0
Jinja2>=3.1.0
tqdm>=4.65.0
PyYAML>=6.0.0
